Added cos_sim functionality
In app.py, json_search(query) now calls a function to perform the cosine sim and return a list of song indices where the first index corresponds to the song with the greatest similarity. If the query (i.e. poem name) isn't in the poems dataset, the function instead returns an empty list.